requirements.txt
discord.py[voice]
gTTS
requests
websockets
sqlalchemy
asyncpg
aiosqlite
redis
python-dotenv
pydantic
